songdo_metr/pyproject.toml
[project]
name = "songdo_metr"
version = "0.1.0"
description = "Default template for PDM package"
authors = [{ name = "Seungbin Yim", email = "ysb06@hotmail.com" }]
dependencies = [
    "torch<=2.3.0",
    "torchvision<=0.18.0",
    "torchaudio<=2.3.0",
    "scikit-learn>=1.5.1",
    "openpyxl>=3.1.5",
    "requests>=2.32.3",
    "geopandas>=1.0.1",
    "tqdm>=4.66.4",
    "tables>=3.9.2",
    "numpy~=1.26.4",
    "pyyaml>=6.0.2",
    "folium>=0.17.0",
    "matplotlib>=3.9.2",
    "mapclassify>=2.8.1",
    "pytest-ordering>=0.6",
]
requires-python = "==3.11.*"
readme = "README.md"
license = { text = "MIT" }

[tool.pdm]
distribution = false

[tool.pdm.dev-dependencies]
dev = [
    "pytest>=8.3.3",
]
